Rename to track_point
