+++ src/Repository/TrackPointRepository.php
@@ -2,21 +2,21 @@
 
 namespace App\Repository;
 
-use App\Entity\LocationPoint;
+use App\Entity\TrackPoint;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
 
 /**
- * @method LocationPoint|null find($id, $lockMode = null, $lockVersion = null)
- * @method LocationPoint|null findOneBy(array $criteria, array $orderBy = null)
- * @method LocationPoint[]    findAll()
- * @method LocationPoint[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ * @method TrackPoint|null find($id, $lockMode = null, $lockVersion = null)
+ * @method TrackPoint|null findOneBy(array $criteria, array $orderBy = null)
+ * @method TrackPoint[]    findAll()
+ * @method TrackPoint[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
-class LocationPointRepository extends ServiceEntityRepository
+class TrackPointRepository extends ServiceEntityRepository
 {
     public function __construct(ManagerRegistry $registry)
     {
-        parent::__construct($registry, LocationPoint::class);
+        parent::__construct($registry, TrackPoint::class);
     }
 
     public function findByBoundingBox(string $neLat, string $neLng, string $swLat, string $swLng)
@@ -27,7 +27,7 @@ class LocationPointRepository extends ServiceEntityRepository
         $nw = "$swLng $neLat";
         // Note start and end points are the same.
         $wkt = "Polygon(($ne, $se, $sw, $nw, $ne))";
-        $sql = "SELECT ST_X(location) AS lng, ST_Y(location) AS lat, timestamp"
+        $sql = "SELECT ST_X(location) AS lng, ST_Y(location) AS lat"
             . " FROM location_point"
             . " WHERE ST_Contains(GeomFromText(:wkt), location)"
             . " ORDER BY RAND()"
