+++ tests/Controller/MapControllerTest.php
@@ -2,12 +2,10 @@
 
 namespace App\Tests;
 
-use App\Repository\LocationPointRepository;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use phpDocumentor\Reflection\Location;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use App\Tests\Controller\ControllerTestBase;
+use App\Repository\TrackPointRepository;
 
-class LocationPointTest extends WebTestCase
+class MapControllerTest extends ControllerTestBase
 {
     /**
      * @dataProvider provideOverlandApiSubmission()
@@ -56,10 +54,10 @@ class LocationPointTest extends WebTestCase
           {"geometry": {"coordinates": [50, 60]}, "properties": {"timestamp": "2021-03-03 18:00:00 +0800"}}
         ]
         }');
-        /** @var LocationPointRepository $lpRepo */
-        $lpRepo = self::$container->get(LocationPointRepository::class);
-        $this->assertSame(2, $lpRepo->count([]));
-        $first = $lpRepo->findOneBy([]);
+        /** @var TrackPointRepository $tpRepo */
+        $tpRepo = self::$container->get(TrackPointRepository::class);
+        $this->assertSame(2, $tpRepo->count([]));
+        $first = $tpRepo->findOneBy([]);
         $this->assertSame('2021-03-03 09:00', $first->getTimestamp()->format('Y-m-d H:i'));
         $this->assertSame(50.0, $first->getLocation()->getLatitude());
         $this->assertSame(40.0, $first->getLocation()->getLongitude());
